netcfg: use proper C++ base type for NetCfgChangeType
NetCfgChangeType represents a type of parameter of the DBus
method NotificationSubscribe, which is a DBus type "u".
According to the specification, this translates to UINT32.

+++ src/netcfg/netcfg-changetype.hpp
@@ -22,7 +22,7 @@
 #include <gdbuspp/glib2/utils.hpp>
 
 
-enum class NetCfgChangeType : std::uint16_t
+enum class NetCfgChangeType : std::uint32_t
 {
     // clang-format off
     UNSET              = 0,         //      0
